+++ changelog.md
@@ -1,5 +1,11 @@
 ﻿## 4.4.0
 - Feat: `/help` command with autocomplete feature
+- Feat: `/gift` commands: `use`, `add`, `remove`, `get`, `accepting`
+- Feat: `/announce` command
+	- announce confirmation button
+	- `announce_gift` as first annouce
+- Update: share functions from `pronouns.js` (`buttonHandler`, `createButton`)
+- Update: add announce personality part where announce_... content is stored
 
 ## 4.3.0
 - Feat: twitter