+++ changelog.md
@@ -1,8 +1,8 @@
 ﻿## 4.4.0
 - Feat: `/help` command with autocomplete feature
 - Feat: `/gift` commands: `use`, `add`, `remove`, `get`, `accepting`
-- Feat: `/anounce` command
-	- announce confirmation `button`
+- Feat: `/announce` command
+	- announce confirmation button
 	- `announce_gift` as first annouce
 - Update: share functions from `pronouns.js` (`buttonHandler`, `createButton`)
 - Update: add announce personality part where announce_... content is stored